PM-360: move provider actions to own file
